// src/app/shared/enums/device-name.enum.ts
export enum DeviceNameEnum {
  angleJet4 = 'valve1',
  angleJet3 = 'valve2',
  angleJet2 = 'valve3',
  angleJet1 = 'valve4',
  verticalJet3 = 'valve5',
  verticalJet2 = 'valve7',
  verticalJet1 = 'valve9',
  foam1 = 'valve10',
  foam2 = 'valve8',
  foam3 = 'valve6',
  ledStrip = 'LED_strip',
  light1 = 'LED3',
  light2 = 'LED2',
  light3 = 'LED1',
  pump1 = 'pump1',
  pump2 = 'pump2',
}